requirements.txt
groq
moviepy
python-dotenv
imageio-ffmpeg